Refactor project structure and update configurations
- Removed `devtools_options.yaml` as it is no longer needed.
- Updated `Makefile` to create Flutter app in the `app` directory.
- Changed `pubspec.yaml` to reflect new workspace structure and removed unnecessary dependencies.
- Added `.gitignore` for the `app` directory to exclude build artifacts and other unnecessary files.
- Introduced localization configuration in `l10n.yaml`.
- Created initial `pubspec.yaml` for the `app` directory.
- Added basic `README.md` for the new Flutter project.
- Implemented error reporting and logging setup.
- Established a new structure for app settings and themes.
- Added initial implementation for home screen and initialization logic.
- Updated tests to reflect new structure and added new test cases.
- Removed unused splash images and old app runner logic.
