+++ app/lib/main.dart
@@ -1,0 +1,3 @@
+import 'package:sizzle_starter/src/feature/initialization/logic/startup.dart';
+
+void main() => startup();
