+++ app/lib/src/core/common/error_reporter/error_reporter.dart
