+++ app/lib/src/core/common/error_reporter/sentry_error_reporter.dart
