+++ app/lib/src/core/common/extensions/string_extension.dart
