+++ app/lib/src/core/common/persisted_entry.dart
