+++ app/lib/src/core/constant/application_config.dart
