+++ app/lib/src/feature/initialization/widget/initialization_failed_app.dart
