+++ app/lib/src/feature/settings/model/app_settings.dart
