+++ app/pubspec.yaml
@@ -1,0 +1,109 @@
+name: sizzle_starter
+description: Production-ready template for Flutter applications.
+# Prevent accidental publishing to pub.dev.
+publish_to: "none"
+
+version: 0.0.1+1
+
+environment:
+  sdk: ">=3.7.0 <4.0.0"
+  flutter: ">=3.29.0 <4.0.0"
+
+resolution: workspace
+
+dependencies:
+  # Flutter SDK
+  flutter:
+    sdk: flutter
+  flutter_localizations:
+    sdk: flutter
+  collection: 1.19.1
+  async: 2.12.0
+  meta: 1.16.0
+  intl: 0.19.0
+
+  # Stream
+  stream_transform: 2.1.1
+  rxdart: 0.28.0
+
+  # BLoC
+  flutter_bloc: 9.1.0
+
+  # Networking
+  http: 1.3.0
+  cronet_http: 1.3.2
+  cupertino_http: 2.0.2
+
+  # Storing data
+  shared_preferences: 2.5.2
+  drift: 2.25.1
+  drift_flutter: 0.2.4
+
+  # Logging
+  stack_trace: 1.12.1
+  sentry_flutter: 8.13.2
+
+  # Packages
+  rest_client:
+    path: ../packages/rest_client
+  app_database:
+    path: ../packages/app_database
+  logger:
+    path: ../packages/logger
+
+  # Utils
+  path: 1.9.1
+  path_provider: 2.1.5
+  clock: 1.1.2
+  fake_async: 1.3.2
+  package_info_plus: 8.3.0
+
+dev_dependencies:
+  # Testing
+  flutter_test:
+    sdk: flutter
+  mockito: 5.4.5
+
+  # Lints
+  sizzle_lints: 2.1.7
+
+  # Utils
+  flutter_gen_runner: 5.10.0
+  flutter_launcher_icons: 0.14.3
+  flutter_native_splash: 2.4.5
+  build_runner: 2.4.15
+
+flutter:
+  uses-material-design: true
+  generate: true
+
+flutter_gen:
+  output: lib/src/core/constant/generated
+  line_length: 80
+
+flutter_launcher_icons:
+  #  image_path: "assets/images/icon-128x128.png"
+  image_path_android: "assets/images/icon-1024x1024.png"
+  image_path_ios: "assets/images/icon-1024x1024.png"
+  android: true # can specify file name here e.g. "ic_launcher"
+  ios: true # can specify file name here e.g. "My-Launcher-Icon"
+  # adaptive_icon_background: "assets/images/christmas-background.png" # only available for Android 8.0 devices and above
+  # adaptive_icon_foreground: "assets/images/icon-foreground-432x432.png" # only available for Android 8.0 devices and above
+  min_sdk_android: 21 # android min sdk min:16, default 21
+  remove_alpha_ios: true
+  background_color_ios: "#ffffff"
+
+  web:
+    generate: true
+    image_path: "assets/images/icon-1024x1024.png"
+    # background_color: "#hexcode"
+    # theme_color: "#hexcode"
+
+  windows:
+    generate: true
+    image_path: "assets/images/icon-1024x1024.png"
+    icon_size: 48 # min:48, max:256, default: 48
+
+  macos:
+    generate: true
+    image_path: "assets/images/icon-1024x1024.png"
