--- devtools_options.yaml
@@ -1,1 +1,0 @@
-extensions:
