--- lib/main.dart
@@ -1,3 +1,0 @@
-import 'package:sizzle_starter/src/feature/initialization/logic/app_runner.dart';
-
-void main() => AppRunner.startup();
